Changes revolving around how the list is loaded

+++ src/coffee/weneed/utils/MathUtil.java
@@ -54,7 +54,7 @@ public class MathUtil {
 	public static int round(double d) {
 		double dAbs = Math.abs(d);
 		int i = (int) dAbs;
-		double result = dAbs - (double) i;
+		double result = dAbs - i;
 		if (result < 0.5) {
 			return d < 0 ? -i : i;
 		} else {
